More factoring out of endpoints in individual files
